updating templatestore instead of using entropy
