+++ src/PAModel/Entropy.cs
@@ -107,12 +107,6 @@ namespace Microsoft.PowerPlatform.Formulas.Tools
         // Key is control name, value is PCFDynamicSchemaForIRRetrieval
         public Dictionary<string, object> PCFDynamicSchemaForIRRetrievalEntry { get; set; } = new Dictionary<string, object>(StringComparer.Ordinal);
 
-        // Key is control name, value is HostType
-        public Dictionary<string, string> HostTypeEntry { get; set; } = new Dictionary<string, string>(StringComparer.Ordinal);
-
-        // Key is control name, value is HostService
-        public Dictionary<string, object> HostServiceEntry { get; set; } = new Dictionary<string, object>(StringComparer.Ordinal);
-
         // Key is control name, value is PCFTemplateDetails
         public Dictionary<string, ControlInfoJson.Template> PCFTemplateEntry { get; set; } = new Dictionary<string, ControlInfoJson.Template>(StringComparer.Ordinal);
